feat: Loading/Error handling
